+++ riddle-renderer-wgpu/src/wgpu_ext/renderable.rs
@@ -1,9 +1,0 @@
-use crate::wgpu_ext::*;
-
-#[doc(hidden)]
-pub struct WGPURenderableDesc<'a> {
-	pub(crate) texture: WGPUTextureHandle,
-	pub(crate) shader: WGPUShaderHandle,
-	pub(crate) verts: &'a [Vertex],
-	pub(crate) indices: &'a [u16],
-}